A SHOCKING BUG fixed
previously, when a tangent is parallel with z axis, that part will get
totally dark no matter how the light and materials are configured. Until
now, with the assistance of VS2015's Graphic Debugger, i find a subtle
but fatal error in INPUT LAYOUT definition : the format of TANGENT is
set to R32G32 ?!! which means that only x and y is fed into the shader.
Damn it!!! It've been a year!!