multi pass post processing - renderTarget switch
It should be available for user to configure which/how the post
processing effect are applied to inital Render Target. because one pass
can achieve an effect, and a single texture can't be bound as shader
input and render target at the same time, multiple backBuffer/Render
Targets are needed. The mechanism to choose which RT to render to is
developed.
But still, how to store a list of post process effect description is
still a problem (perhaps an array of Variant?)